Generate fed templates (#208)
* WIP

* Add beginnings of RIOT Makefile

* Implement LFC templates for RIOT (#210)

* Implement LFC templates for RIOT

* Remove makefile from coap-federated example

* Refactor string formatting a little

* Fix RIOT makefile generation

---------


* Add `generate.sh` file to generated fed templates (#215)

* Generate generate.sh files to invoke lfc from a federate

* Spotless

---------


* generate.sh -> run_lfc.sh

* Split generated CMake into Include.cmake and CMakeLists.txt

* Fix shell script

* Sanitize input to lf_build_generated_code

* Updates

* Fix riot test

* Fix riot example

* Run formatter

* FIx merge

* Debug CI

* correct path to enc

* Formatting

* Test gen-fed-templates in CI

* CI

* CI

* Dont do RIOT yet because of CI issue

* CI

* Formating

* Delete mistakenly committed files

---------

+++ src/serialization.c
@@ -38,7 +38,7 @@ lf_ret_t deserialize_payload_default(void *user_struct, const unsigned char *msg
   return LF_OK;
 }
 
-ssize_t serialize_payload_default(const void *user_struct, size_t user_struct_size, unsigned char *msg_buf) {
+int serialize_payload_default(const void *user_struct, size_t user_struct_size, unsigned char *msg_buf) {
   if (user_struct_size > SERIALIZATION_MAX_PAYLOAD_SIZE) {
     return -1;
   }
